requirements.txt
annotated-types==0.7.0
anyio==4.8.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Cython==0.29.36
docx2txt==0.8
fastapi[standard]==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
openai==1.75.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
psycopg2==2.9.10
alembic==1.15.1
bleach==6.2.0
pydantic[email]
asyncio==3.4.3
clerk-backend-api==2.0.2